Add playlist visibility and invitations
Introduce playlist visibility (Private/Unlisted/Public) and a full invitations flow. Replaces Playlist.IsPublic with Playlist.Visibility and adds PlaylistVisibility enum plus EF migration to add the column and create PlaylistInvitations table. Add PlaylistInvitation entity, DataContext configuration, AutoMapper mappings and DTO (PlaylistInvitationResponse). Implement IPlaylistInvitationService/PlaylistInvitationService and PlaylistInvitationsController (invite, accept, decline, cancel, incoming/outgoing/count). Update PlaylistsController to enforce visibility, add save/unsave and visibility update endpoints, and move ownership/admin checks into services. Wire up DI in Program. Update related models (PlaylistResponse, CreatePlaylistRequest, friendship/relationship status types) and many frontend files (models, pages, components, stores and thunks) to support visibility and invitations. These changes enable shareable/unlisted playlists and a notification/invitation UX while preventing private playlist leakage.

+++ backend/Services/FriendshipService.cs
@@ -81,7 +81,7 @@ namespace SongAppApi.Services
                         {
                             SenderId = senderId,
                             ReceiverId = receiverId,
-                            Status = (int)FriendshipStatus.Pending,
+                            Status = FriendshipStatus.Pending,
                             CreatedAt = DateTime.UtcNow,
                         };
                 }
@@ -253,7 +253,7 @@ namespace SongAppApi.Services
 
             return new RelationshipStatusResponse
             {
-                Status = (int)row.Status,
+                Status = row.Status,
                 IsCurrentUserSender = row.SenderId == currentGuid,
                 IsSelf = false,
             };
